Fix login always failing against dbo.USERS
Two bugs, both only reproducible against the real SQL Server backend
(masked by the InMemory-only test suite):

- ExternalUserService.ValidateAsync used string.ToUpperInvariant()
  inside the LINQ query; EF Core's SQL Server provider can't translate
  it, so every login attempt threw and was swallowed by the generic
  catch. Replaced with .ToUpper(), which EF Core maps to SQL UPPER().
- AppUser.Id/IdUserProfile were typed int/int?, but the real dbo.USERS
  columns are smallint, causing an InvalidCastException reading the
  data reader. Retyped to short/short?, matching the existing
  Agent/Player smallint mappings.

Verified by rebuilding to an isolated output path, running against
the live DGSDataTest connection, and reproducing a login via curl:
now returns a real Identity auth cookie instead of the generic error.

+++ Models/AppUser.cs
@@ -6,11 +6,11 @@ namespace DgsTool.Models;
 /// </summary>
 public class AppUser
 {
-    public int Id { get; set; }
+    public short Id { get; set; }
     public string LoginName { get; set; } = string.Empty;
     public string Password { get; set; } = string.Empty;
     public string? Name { get; set; }
     public bool Status { get; set; }
     public byte Type { get; set; }
-    public int? IdUserProfile { get; set; }
+    public short? IdUserProfile { get; set; }
 }
